+++ NEWS.md
@@ -1,3 +1,5 @@
+# TemporalHazard 1.1.0.9000 (development version)
+
 # TemporalHazard 1.1.0
 
 ## New features